backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
sqlalchemy==2.0.35
python-multipart==0.0.12

# PDF Processing
PyMuPDF==1.24.10
Pillow==10.4.0
numpy==1.26.4

# AI/ML
google-generativeai==0.8.3
google-search-results==2.4.2  # SerpAPI for web search
deepgram-sdk==3.2.7  # Real-time speech-to-text

# Storage & Database
supabase==2.9.0

# Vector Database
pinecone-client==5.0.1

# Additional utilities
aiofiles==24.1.0
